requirements.txt
bcrypt==4.0.1
Flask==2.2.3
mysql-connector-python==8.0.32
python-dotenv==1.0.0
stripe==5.4.0
qrcode==7.4.2
waitress==2.1.2
sqlalchemy==2.0.12
sendgrid==6.10.0